Gutted useless junk and reorganized methods.
I cut out all of the extra key requiring things. The app
won't work as long as JumpstartAuth doesn't use
rather than the regular gets, but I don't think it will be a
problem since I got the issue looked at by a collaborator.
The next challenge will be to make the large tweets reply to
themselves rather than post * amount of standalone tweets.
I also found the order of the methods to be confusing, so I moved
them around a bit.